+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.2.0
+## Unreleased (proposed: 2.2.0)
 Adds support to API300 HPE Synergy Image Streamer resources:
   - image_streamer_artifact_bundle
   - image_streamer_os_build_plan
@@ -6,6 +6,9 @@ Adds support to API300 HPE Synergy Image Streamer resources:
 Bug Fixes:
 - [#98](https://github.com/HewlettPackard/oneview-chef/issues/98) Fix get_diff for comparisons of alike data
 - [#145](https://github.com/HewlettPackard/oneview-chef/issues/145) Show diff on log statement before actual update
+
+Enhancements:
+- [#218](https://github.com/HewlettPackard/oneview-chef/issues/218) connection_template resource does not support NetworkSet, FCoENetwork and FCNetwork
 
 ## 2.1.0
 - [#162](https://github.com/HewlettPackard/oneview-chef/issues/162) Add server_profile_template property to oneview_server_profile